Make Qt::Initialization a C++11 class enum
And declare Qt::Uninitialized to be the same value because we're not
interested in the scoping rules of C++11 class enums. We're only
interested in avoiding the cast from Qt::Uninitialized to an integer.
That is to avoid the mistaken:

    QVector<Custom> vector(5, Qt::Uninitialized);

which is actually implicitly:

    QVector<Custom> vector(5, Custom(Qt::Uninitialized));

and likely not what the developer wanted.

Change-Id: I27eaacb532114dd188c4ffff13d374eb698bfbab
Reviewed-by: Marc Mutz <marc.mutz@kdab.com>