cmake: Generate INTERFACE targets for include-only modules.
Require CMake 3.0 if an attempt is made to use a cmake file containing
an INTERFACE library.

If the user is using a CMake version older than 3.0, then exclude INTERFACE
libraries from dependencies of Qt modules.  The Qt CI system is running
CMake versions as old as 2.8.11, which makes that the current minimum version.

The only header-only module existing so far is the QtUiPlugin module, which
has been split out from the QtDesigner module.  If using CMake 2.8, the
forwarding headers in the QtDesigner module will be used, and the effect
of the split out library will not be seen.  If using CMake 3.0, the
split out library is listed as a dependency and its transitive usage
requirements such as the QT_UIPLUGIN_LIB definition are made available.

Change-Id: Iecee3bbc440842dca27dc067f2a31e3526efa01b
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>