ios: resolve m_assetUrl already in QIOSFileEngineAssetsLibrary::setFileName()
Resolve m_assetUrl already when setting file name. The variable will be
used several places in patches that follows.

At the same time, change the logic to be more robust to work around QDir
removing slashes (both single a double) after the scheme. In the end, what
matters is that we still recognize the file name as an asset url, and that we can
restore the original url based on the hash-tag contained inside the file name.

Change-Id: I988c6a73b2484e46d63917b442c13aa5a3666787
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@theqtcompany.com>