Fix printing page margins wrong values on windows
There is an error. The page margins have wrong values because of float overrange. Huge value is assigned to float and then divided by a 'multiplier'. The solution is first to divide and rhen assigne it to float.

This error was resolved in 5.3.0 or 5.3.1 version (don't remember) and then the 5.3.2 introduced it again!

This error was reported in many many bug issues.