Doc: Excluded qdoc files that caused unnecessary Qt Creator warnings.
The customcompleter and textcodes are widget examples, but they end up
in the Qt Core exampledirs boundary because of a reference to the
plugandpaint example in the docs. This resulted in a couple of wrong
entries being written into the examples-manifest.xml, which is used
by Qt Creator.

This change explicitly exludes the qdoc pages for the two examples
so that qdoc doesn't add the corresponding entries into
examples-manifest.xml.

Task-number: QTBUG-41996
Change-Id: I0e95b6d4d93e0ce18f5b34e5034b279598b4924f
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>
Reviewed-by: Martin Smith <martin.smith@digia.com>